// packages/blueprints/src/generated/studio-config.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run "yarn generate-schema-types" to regenerate this file.
 */

export interface StudioConfig {
	visionMixerType: 'atem' | 'vmix'
	atemSources: {
		[k: string]: InputConfig
	}
	atemOutputs: {
		[k: string]: OutputConfig
	}
	vmixSources: {
		[k: string]: VmixInputConfig
	}
	sisyfosSources: {
		[k: string]: SiyfosSourceConfig
	}
	previewRenderer: string
	casparcgLatency: number
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "".
 */
export interface InputConfig {
	input: number
	type: string
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "".
 */
export interface OutputConfig {
	output: number
	source: number
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "".
 */
export interface VmixInputConfig {
	input: number
	type: string
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "".
 */
export interface SiyfosSourceConfig {
	source: number
	type: string
}